Move backoffice frontend DI out of config folder
