--- src/apps/backoffice/frontend/config/dependency-injection/Shared/application.yaml
@@ -1,35 +1,0 @@
-services:
-  Shared.Logger:
-    class: ../../../../../../Contexts/Shared/infrastructure/WinstonLogger
-    arguments: []
-
-  Shared.MongoConfig:
-    factory:
-      class: ../../../../../../Contexts/Mooc/Shared/infrastructure/persistence/mongo/MongoConfigFactory
-      method: 'createConfig'
-
-  Shared.ConnectionManager:
-    factory:
-      class: ../../../../../../Contexts/Shared/infrastructure/persistence/mongo/MongoClientFactory
-      method: 'createClient'
-    arguments: ['mooc', '@Shared.MongoConfig']
-
-  Shared.EventBus:
-    class: ../../../../../../Contexts/Shared/infrastructure/EventBus/InMemoryAsyncEventBus
-    arguments: []
-
-  Shared.QueryBus:
-    class: ../../../../../../Contexts/Shared/infrastructure/QueryBus/InMemoryQueryBus
-    arguments: ['@Shared.QueryHandlersInformation']
-
-  Shared.CommandBus:
-    class: ../../../../../../Contexts/Shared/infrastructure/CommandBus/InMemoryCommandBus
-    arguments: ['@Shared.CommandHandlersInformation']
-
-  Shared.QueryHandlersInformation:
-    class: ../../../../../../Contexts/Shared/infrastructure/QueryBus/QueryHandlersInformation
-    arguments: ['!tagged queryHandler']
-
-  Shared.CommandHandlersInformation:
-    class: ../../../../../../Contexts/Shared/infrastructure/CommandBus/CommandHandlersInformation
-    arguments: ['!tagged commandHandler']
