+++ src/apps/backoffice/frontend/dependency-injection/application.yaml
