+++ src/apps/backoffice/frontend/dependency-injection/application_production.yaml
