+++ src/apps/backoffice/frontend/dependency-injection/application_staging.yaml
