+++ src/apps/backoffice/frontend/dependency-injection/application_test.yaml
