+++ src/apps/backoffice/frontend/dependency-injection/index.ts
