+++ src/apps/backoffice/frontend/seed.ts
@@ -1,7 +1,7 @@
 import { CoursesCounterRepository } from '../../../Contexts/Mooc/CoursesCounter/domain/CoursesCounterRepository';
 import { CoursesCounter } from '../../../Contexts/Mooc/CoursesCounter/domain/CoursesCounter';
 import { CoursesCounterId } from '../../../Contexts/Mooc/CoursesCounter/domain/CoursesCounterId';
-import container from './config/dependency-injection';
+import container from './dependency-injection';
 
 export async function seed() {
   const repository: CoursesCounterRepository = container.get('Mooc.coursesCounter.CoursesCounterRepository');
